Début gestion FileDefinition + Documentation + DummyData

+++ FinanceManager/Import/Mapper/CsvExtensions.cs
@@ -12,7 +12,8 @@ namespace FinanceManager.Import.Extension
             //else if (!string.IsNullOrEmpty(item.CC)) { ft.AccountNumber = item.CC; }
             //else if (!string.IsNullOrEmpty(accountNumber)) { ft.AccountNumber = accountNumber; }
             //else { ft.AccountNumber = item.}
-            ft.AccountNumber = item.Compte.ToInt();
+            ft.AccountNumber = item.Compte.ToLong();
+            //ft.AccountNumber = item.Compte.Replace("*","").Replace("0","").ToLong();
             ft.TimeStamp = item.TransactionDate.ToDate();
             ft.TransactionType = item.TransactionType.AdjustSpecialCaracters().ToUpper();
             ft.Description = string.Format("{0} {1}", item.Description, item.SubDescription).Trim();
